requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
requests==2.27.1
six==1.15.0
urllib3==1.26.8
cffi==1.14.4
cryptography==3.3.2
pycparser==2.20
PyMySQL==0.10.1
python-dotenv==0.15.0
six==1.15.0

